updates label style for LevelSelect in vizbuilder updates composePropertyName to show optional combinations moves formatting code to its own file (related to #136)
