+++ packages/vizbuilder/src/helpers/formatting.js
@@ -1,0 +1,40 @@
+export const PROPNAMESTYLES = {
+  LVL: 1,
+  HIE: 2,
+  HIELVL: 3,
+  DIM: 4,
+  DIMLVL: 5,
+  DIMHIE: 6,
+  DIMHIELVL: 7
+};
+
+/**
+ * For a drilldown, generates a standard name format to use in selectors.
+ * @param {Level|Measure} item A Level or Measure object
+ * @param {number?} style A combination available from the PROPNAMESTYLES enum
+ * @returns {string}
+ */
+export function composePropertyName(item, style = PROPNAMESTYLES.DIMHIELVL) {
+  let txt = style & PROPNAMESTYLES.LVL ? item.name : "";
+  if ("hierarchy" in item) {
+    const hname = item.hierarchy.name;
+    const dname = item.hierarchy.dimension.name;
+    if (style & PROPNAMESTYLES.HIE && hname !== item.name && hname !== dname) {
+      txt = `${item.hierarchy.name} › ${txt}`;
+    }
+    if (style & PROPNAMESTYLES.DIM && dname !== item.name) {
+      txt = `${dname} › ${txt}`;
+    }
+  }
+  return txt;
+}
+
+/**
+ * Generates a string that can be used as index to sort elements.
+ * @param {string} string The string to slice
+ * @returns {string}
+ */
+export function sortSlice(string) {
+  string = `${string}`.replace(/\W/g, "").toLowerCase();
+  return `${string.slice(0, 5)}-----`.slice(0, 6);
+}
